Refactored components to a single list
We've refactored Places and Passenger components to a single
reusable component so we can solve CodeClimate issues.
