+++ FlightReservation/src/components/booking/consts/ListPassengers.js
@@ -1,0 +1,15 @@
+const listPassengers = [
+  {label: 'Select number:', value: null},
+  {label: '1', value: '1 passenger'},
+  {label: '2', value: '2 passengers'},
+  {label: '3', value: '3 passengers'},
+  {label: '4', value: '4 passengers'},
+  {label: '5', value: '5 passengers'},
+  {label: '6', value: '6 passengers'},
+  {label: '7', value: '7 passengers'},
+  {label: '8', value: '8 passengers'},
+  {label: '9', value: '9 passengers'},
+  {label: '10', value: '10 passengers'},
+];
+
+export default listPassengers;
