+++ FlightReservation/src/components/booking/consts/ListPlaces.js
@@ -1,0 +1,12 @@
+const listPlaces = [
+  {label: 'Select location:', value: null},
+  {label: 'Ciudad de Mexico', value: 'Ciudad de Mexico'},
+  {label: 'Aguascalientes', value: 'Aguascalientes'},
+  {label: 'Colima', value: 'Colima'},
+  {label: 'Sinaloa', value: 'Sinaloa'},
+  {label: 'Zacatecas', value: 'Zacatecas'},
+  {label: 'Chihuahua', value: 'Chihuahua'},
+  {label: 'Oaxaca', value: 'Oaxaca'},
+];
+
+export default listPlaces;
